+++ tests/test_ruuvitag_sensor.py
@@ -14,8 +14,8 @@ class TestRuuviTagSensor(TestCase):
         data = '043E2A0201030157168974A5F41E0201060303AAFE1616AAFE10EE037275752E76692F23416A7759414D4663CD'
         return data[26:]
 
-    @patch('ruuvitag_sensor.adaptors.nix_hci.BleCommunicationNix.get_data', get_data)
-    @patch('ruuvitag_sensor.adaptors.dummy.BleCommunicationDummy.get_data', get_data)
+    @patch('ruuvitag_sensor.adapters.nix_hci.BleCommunicationNix.get_data', get_data)
+    @patch('ruuvitag_sensor.adapters.dummy.BleCommunicationDummy.get_data', get_data)
     def test_tag_update_is_valid(self):
         tag = RuuviTag('48:2C:6A:1E:59:3D')
 
@@ -55,17 +55,17 @@ class TestRuuviTagSensor(TestCase):
         for data in datas:
             yield data
 
-    @patch('ruuvitag_sensor.adaptors.dummy.BleCommunicationDummy.get_datas',
+    @patch('ruuvitag_sensor.adapters.dummy.BleCommunicationDummy.get_datas',
            get_datas)
-    @patch('ruuvitag_sensor.adaptors.nix_hci.BleCommunicationNix.get_datas',
+    @patch('ruuvitag_sensor.adapters.nix_hci.BleCommunicationNix.get_datas',
            get_datas)
     def test_find_tags(self):
         tags = RuuviTagSensor.find_ruuvitags()
         self.assertEqual(7, len(tags))
 
-    @patch('ruuvitag_sensor.adaptors.dummy.BleCommunicationDummy.get_datas',
+    @patch('ruuvitag_sensor.adapters.dummy.BleCommunicationDummy.get_datas',
            get_datas)
-    @patch('ruuvitag_sensor.adaptors.nix_hci.BleCommunicationNix.get_datas',
+    @patch('ruuvitag_sensor.adapters.nix_hci.BleCommunicationNix.get_datas',
            get_datas)
     def test_get_data_for_sensors(self):
         macs = ['CC:2C:6A:1E:59:3D', 'DD:2C:6A:1E:59:3D', 'EE:2C:6A:1E:59:3D']
@@ -77,18 +77,18 @@ class TestRuuviTagSensor(TestCase):
         self.assertTrue(data['EE:2C:6A:1E:59:3D']['temperature'] == 25.12)
         self.assertTrue(data['EE:2C:6A:1E:59:3D']['identifier'] == '0')
 
-    @patch('ruuvitag_sensor.adaptors.dummy.BleCommunicationDummy.get_datas',
+    @patch('ruuvitag_sensor.adapters.dummy.BleCommunicationDummy.get_datas',
            get_datas)
-    @patch('ruuvitag_sensor.adaptors.nix_hci.BleCommunicationNix.get_datas',
+    @patch('ruuvitag_sensor.adapters.nix_hci.BleCommunicationNix.get_datas',
            get_datas)
     def test_get_datas(self):
         datas = []
         RuuviTagSensor.get_datas(datas.append)
         self.assertEqual(7, len(datas))
 
-    @patch('ruuvitag_sensor.adaptors.dummy.BleCommunicationDummy.get_datas',
+    @patch('ruuvitag_sensor.adapters.dummy.BleCommunicationDummy.get_datas',
            get_datas)
-    @patch('ruuvitag_sensor.adaptors.nix_hci.BleCommunicationNix.get_datas',
+    @patch('ruuvitag_sensor.adapters.nix_hci.BleCommunicationNix.get_datas',
            get_datas)
     def test_get_datas_with_macs(self):
         datas = []