+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 
 ### [Unreleased]
+* Fix usage of bt_device parameter when opening the BLE connection
 
 ## [0.13.0] - 2019-07-01
 * Fix hcitool subprocess closing