+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Changelog
 
 ### [Unreleased]
+
+## [1.0.0] - 2020-03-21
 * Fix usage of bt_device parameter when opening the BLE connection
 * Refactor adapters into own modules
 * Fix blacklisting of non Ruuvitag devices